UI Manager singleton, fixed minor bugs on Main Menu and created settings page
Fixed a few smaller things on the main menu, created the settings menu (still need to program the logic) and created UI Manager with singletons
