--- Assets/Scripts/CursorManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 19767bb0e8ad8cc4eac31a2734af7137
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
