--- Assets/Scripts/ScenesSwitcher.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 85accc55708059a46acd0c55f7a5f9ec
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
