--- Assets/Scripts/SwapSprite.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e295dfd8bfac06444b7a3adcec990d8b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
