+++ Assets/Scripts/UI/ScenesSwitcher.cs
@@ -1,0 +1,14 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.SceneManagement;
+
+public class ScenesSwitcher : MonoBehaviour
+{
+    SceneManager sceneManager;
+
+    public void LoadNewScene(int sceneNumber)
+    {
+        UnityEngine.SceneManagement.SceneManager.LoadScene(sceneNumber);
+    }
+}
