+++ Assets/Scripts/UI/SwapSprite.cs
@@ -1,0 +1,20 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.UI;
+using UnityEngine.EventSystems;
+
+public class SwapSprite : MonoBehaviour, IPointerEnterHandler, IPointerExitHandler
+{
+    public GameObject HoverPanel;
+
+    public void OnPointerEnter(PointerEventData eventData)
+    {
+        HoverPanel.SetActive(true);
+    }
+
+    public void OnPointerExit(PointerEventData eventData)
+    {
+        HoverPanel.SetActive(false);
+    }
+}
