+++ Assets/User Interface/Icon.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4691999b8c1c3604aa700767857f4590
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
